2019 intcode int64 support

+++ 2019/day05/main.go
@@ -13,8 +13,8 @@ func main() {
 
 	filepath := "day05/input.txt"
 	ic := intcode.GetIntcode(filepath)
-	res := intcode.RunWithInputs(slices.Clone(ic[:]), []int{1})
+	res := intcode.RunWithInputs(slices.Clone(ic[:]), []int64{1})
 	fmt.Printf("Part 1: %v\n", res)
-	res = intcode.RunWithInputs(slices.Clone(ic[:]), []int{5})
+	res = intcode.RunWithInputs(slices.Clone(ic[:]), []int64{5})
 	fmt.Printf("Part 2: %v\n", res)
 }
